Refactor portfolio data handling and add unsaved changes banner component
